+++ commands/test.js
@@ -1,11 +1,12 @@
 ﻿var Discord = require("discord.js"),
-	request = require("request");
+	request = require("request"),
+	Config = require("../config.json");
 
 module.exports = {
 	command: "test",
 	catalog: 1,
 	server: [],
-	user: ["210125258034905089"],
+	user: [Config.discord_options.bot_owner],
 	role: [],
 	hidden: true,
 	description: "",
@@ -24,12 +25,9 @@ module.exports = {
 						cover = body.results[0].coverUrl,
 						releaseUrl = body.results[0].urls[0];
 					
-					
-					
 					cover = cover.replace(/ /g, "%20");
 					if(releaseUrl) releaseUrl = releaseUrl.replace(/ /g, "%20");
 					
-					//console.log(body);
 					const embed = new Discord.RichEmbed()
 						.setTitle("🆕 Latest Release")
 						.addField(`${song}`,`${artist}`)
@@ -38,8 +36,6 @@ module.exports = {
 					if(releaseUrl) embed.setURL(`${releaseUrl}`);
 					if(!releaseUrl) embed.addField("Links","None");
 						
-					/*console.log({embed});
-					console.log(song + "," + artist + "," + cover + "," + releaseUrl);*/
 					message.channel.send({embed});
 					
 				}else console.log(error);